+++ docs/pages/_app.js
@@ -9,10 +9,6 @@ import PropTypes from 'prop-types';
 import acceptLanguage from 'accept-language';
 import { create } from 'jss';
 import jssRtl from 'jss-rtl';
-import { StyleSheetManager } from 'styled-components';
-import { CacheProvider } from '@emotion/react';
-import createCache from '@emotion/cache';
-import rtlPlugin from 'stylis-plugin-rtl';
 import { useRouter } from 'next/router';
 import { StylesProvider, jssPreset } from '@material-ui/styles';
 import pages from 'docs/src/pages';
@@ -20,11 +16,13 @@ import initRedux from 'docs/src/modules/redux/initRedux';
 import PageContext from 'docs/src/modules/components/PageContext';
 import GoogleAnalytics from 'docs/src/modules/components/GoogleAnalytics';
 import loadScript from 'docs/src/modules/utils/loadScript';
-import RtlContext from 'docs/src/modules/utils/RtlContext';
 import { ThemeProvider } from 'docs/src/modules/components/ThemeContext';
 import { pathnameToLanguage, getCookie } from 'docs/src/modules/utils/helpers';
 import { ACTION_TYPES, CODE_VARIANTS, LANGUAGES } from 'docs/src/modules/constants';
 import { useUserLanguage } from 'docs/src/modules/utils/i18n';
+import StyledEngineProvider, { cacheLtr } from 'docs/src/modules/utils/StyledEngineProvider';
+
+export { cacheLtr };
 
 // Configure JSS
 const jss = create({
@@ -281,17 +279,6 @@ function findActivePage(currentPages, pathname) {
   return activePage;
 }
 
-// Cache for the ltr version of the styles
-export const cacheLtr = createCache({ key: 'css' });
-cacheLtr.compat = true;
-
-// Cache for the rtl version of the styles
-const cacheRtl = createCache({
-  key: 'rtl',
-  stylisPlugins: [rtlPlugin],
-});
-cacheRtl.compat = true;
-
 function AppWrapper(props) {
   const { children, pageProps } = props;
 
@@ -333,17 +320,13 @@ function AppWrapper(props) {
         ))}
       </NextHead>
       <ReduxProvider store={redux}>
-        <RtlContext.Provider value={rtlContextValue}>
-          <PageContext.Provider value={{ activePage, pages, versions: pageProps.versions }}>
-            <StyleSheetManager stylisPlugins={rtl ? [rtlPlugin] : []}>
-              <CacheProvider value={rtl ? cacheRtl : cacheLtr}>
-                <StylesProvider jss={jss}>
-                  <ThemeProvider>{children}</ThemeProvider>
-                </StylesProvider>
-              </CacheProvider>
-            </StyleSheetManager>
-          </PageContext.Provider>
-        </RtlContext.Provider>
+        <PageContext.Provider value={{ activePage, pages, versions: pageProps.versions }}>
+          <StylesProvider jss={jss}>
+            <ThemeProvider>
+              <StyledEngineProvider>{children}</StyledEngineProvider>
+            </ThemeProvider>
+          </StylesProvider>
+        </PageContext.Provider>
         <LanguageNegotiation />
         <Analytics />
       </ReduxProvider>
