--- docs/src/modules/utils/RtlContext.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-const RltContext = React.createContext({
-  rtl: false,
-  setRtl: () => {},
-});
-
-export default RltContext;
